enhancement(Updates): Move getUsers() and specify target roles
